SDK-75 extract validation of custom enc methods into separate file
