+++ lib/in-crypt.js
@@ -3,12 +3,12 @@ const util = require('util');
 const { DEFAULT_VERSION } = require('./secret-key-accessor');
 const { StorageCryptoError, StorageClientError } = require('./errors');
 const { isValid, getErrorMessage } = require('./validation/utils');
-const { reflect } = require('./utils');
 const {
   CustomEncryptionConfigsIO,
   CUSTOM_ENCRYPTION_ERROR_MESSAGE_ENC,
   CUSTOM_ENCRYPTION_ERROR_MESSAGE_DEC,
 } = require('./validation/custom-encryption-configs');
+const { validateCustomEncryption } = require('./validation/custom-encryption-configs-deep');
 
 /**
  * @typedef {import('./validation/custom-encryption-configs').CustomEncryptionConfig} CustomEncryptionConfig
@@ -43,7 +43,11 @@ class InCrypt {
     }
 
     if (this.customEncryption !== null) {
-      await this.validateCustomEncryption();
+      if (this.secretKeyAccessor === undefined) {
+        throw new StorageCryptoError(CUSTOM_ENCRYPTION_ERROR_MESSAGE_NO_SKA);
+      }
+      const secretData = await this.secretKeyAccessor.getSecrets();
+      await validateCustomEncryption(secretData, Object.values(this.customEncryption));
     }
   }
 
@@ -74,57 +78,6 @@ class InCrypt {
     if (current) {
       this.currentCustomEncryptionVersion = this.packCustomEncryptionVersion(current.version);
     }
-  }
-
-  async validateCustomEncryption() {
-    if (this.secretKeyAccessor === undefined) {
-      throw new StorageCryptoError(CUSTOM_ENCRYPTION_ERROR_MESSAGE_NO_SKA);
-    }
-
-    const secretData = await this.secretKeyAccessor.getSecrets();
-    const secrets = secretData.secrets.filter((s) => s.isForCustomEncryption);
-    if (secrets.length === 0) {
-      throw new StorageCryptoError('No secret for Custom Encryption');
-    }
-
-    const results = await Promise.all(
-      secrets.map((secret) => Object.values(this.customEncryption).map(async (config) => {
-        const context = `[Custom enc version: ${config.version}; Secret version: ${secret.version}]`;
-        const plaintext = 'incountry';
-        try {
-          const enc = await config.encrypt(plaintext, secret.secret, secret.version);
-          if (typeof enc !== 'string') {
-            throw new StorageCryptoError(`${context} ${CUSTOM_ENCRYPTION_ERROR_MESSAGE_ENC}. Got ${typeof enc}`);
-          }
-        } catch (e) {
-          throw new StorageCryptoError(`${context} ${e.message}`);
-        }
-
-        try {
-          const enc = await config.encrypt(plaintext, secret.secret, secret.version);
-          const dec = await config.decrypt(enc, secret.secret, secret.version);
-
-          if (typeof dec !== 'string') {
-            throw new StorageCryptoError(`${context} ${CUSTOM_ENCRYPTION_ERROR_MESSAGE_DEC}. Got ${typeof dec}`);
-          }
-
-          if (dec !== plaintext) {
-            throw new StorageCryptoError(`${context} decrypted data doesn't match the original input`);
-          }
-        } catch (e) {
-          throw new StorageCryptoError(`${context} ${e.message}`);
-        }
-      }))
-        .flat()
-        .map(reflect),
-    );
-
-    const errors = results.filter((result) => result.status === 'rejected').map((result) => result.reason);
-    if (errors.length) {
-      return Promise.reject(errors);
-    }
-
-    return results;
   }
 
   async getCurrentSecretVersion() {
